requirements.txt
opencv-python
flask
matplotlib
os
flask_sqlalchemy
imageio
numpy
copy
paddlepaddle